--- src/App.tsx
@@ -1,48 +1,0 @@
-import React from 'react';
-import Header from './Header'
-import Params from './Params'
-import Calendar from './Calendar'
-import './App.css';
-import { createStyles, Theme, makeStyles } from '@material-ui/core/styles';
-import CssBaseline from '@material-ui/core/CssBaseline';
-
-const drawerWidth = 240;
-const useStyles = makeStyles((theme: Theme) => 
-  createStyles({
-    root: {
-      display: 'flex',
-    },
-    appBar: {
-      zIndex: theme.zIndex.drawer + 1,
-    },
-    drawer: {
-      width: drawerWidth,
-      flexShrink: 0,
-    },
-    drawerPaper: {
-      width: drawerWidth,
-    },
-    drawerContainer: {
-      overflow: 'auto',
-    },
-    content: {
-      flexGrow: 1,
-      padding: theme.spacing(3),
-    },
-  })
-);
-
-function App() {
-  const classes = useStyles();
-
-  return (
-    <div className={classes.root}>
-      <CssBaseline />
-      <Header classes={classes} />
-      <Params classes={classes} />
-      <Calendar classes={classes} />
-    </div>
-  );
-}
-
-export default App;
