+++ src/Styles/Styles.tsx
@@ -1,0 +1,30 @@
+import { createStyles, Theme, makeStyles } from '@material-ui/core/styles';
+
+export default function getStyles(): any {
+    const drawerWidth = 240;
+    return makeStyles((theme: Theme) =>
+        createStyles({
+            root: {
+                display: 'flex',
+            },
+            appBar: {
+                zIndex: theme.zIndex.drawer + 1,
+                'background-color': '#ff5e00',
+            },
+            drawer: {
+                width: drawerWidth,
+                flexShrink: 0,
+            },
+            drawerPaper: {
+                width: drawerWidth,
+            },
+            drawerContainer: {
+                overflow: 'auto',
+            },
+            content: {
+                flexGrow: 1,
+                padding: theme.spacing(3),
+            },
+        })
+    )();
+}
